Moved appropriate tests into game.rs, allowing for private 'players' field
